requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.30
aiosqlite>=0.19.0
pydantic>=2.5.0
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.9
email-validator>=2.2.0
